refactor(static-export): split engine into focused modules
Addresses PR #225 should-fix #5: break up the 1.5k-line monolith into
discrete files so each concern is easy to find and review.

  lib/apps/static-export/
    config.ts          — getExportConfig / saveExportConfig + last-job
                         persistence (getLastExportJob / saveLastExportJob)
    document.ts        — buildDocument, renderPageBody, escapeHtml, SEO
                         helpers, Swiper constants, slider boot script,
                         interactions boot script, collectInteractions,
                         layerTreeContains
    paths.ts           — computeOutputKey (S3-friendly clean URLs)
    resolver.ts        — resolvePages async-generator, renderResolved,
                         LocaleContext, buildTranslationsMap
    asset-bundler.ts   — collectPublicAssets + collectSupabaseAssets
                         (with /a/<hash> proxy-URL resolution)
    writers/types.ts   — OutputFile, Writer, MIME helpers
    writers/local.ts   — createLocalWriter
    writers/s3.ts      — createS3Writer (dynamic @aws-sdk import)
    writers/github.ts  — createGithubWriter (clone / commit / push)
    engine.ts          — exportSite orchestrator
    index.ts           — thin re-export of the public API

No behavioural change. Largest file is now 458 lines (document.ts —
dominated by the inline JS boot scripts). External call sites continue
to import from `@/lib/apps/static-export`.
